Update the optional tools and add command-line commnads for them. Update README
